+++ TFLM_PersonDetection/tensorflow/lite/micro/tools/make/downloads/xi_annlib_vision_p6/lib_common/cnnrt.cc
@@ -711,7 +711,7 @@ void dma_2d_loc2sys_dyn_ch(int ch,
                            void *src, void *dst, size_t d1_cnt, int d2_src_pitch, int d2_dst_pitch, size_t d2_cnt)
 {
     unsigned added;
-    if (d1_cnt == d2_src_pitch && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch && (int)d1_cnt == d2_dst_pitch) {
         added = add_2d_desc(ch,
                             src, dst, d1_cnt * d2_cnt, d1_cnt * d2_cnt, d1_cnt * d2_cnt, 1,
                             /*added*/ 0, /*isSys2Loc*/ 0, /*straddles*/ 0);
@@ -729,7 +729,7 @@ void dma_2d_sys2loc_dyn_ch(int ch,
                            void *src, void *dst, size_t d1_cnt, int d2_src_pitch, int d2_dst_pitch, size_t d2_cnt)
 {
     unsigned added;
-    if (d1_cnt == d2_src_pitch && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch && (int)d1_cnt == d2_dst_pitch) {
         added = add_2d_desc(ch,
                             src, dst, d1_cnt * d2_cnt, d1_cnt * d2_cnt, d1_cnt * d2_cnt, 1,
                             /*added*/ 0, /*isSys2Loc*/ 1, /*straddles*/ 0);
@@ -821,12 +821,12 @@ void dma_3d_loc2sys_dyn_ch(int ch,
                            int d3_src_pitch, int d3_dst_pitch, size_t d3_cnt)
 {
     /* Fold dimensions if possible */
-    if (d1_cnt == d2_src_pitch
-        && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch
+        && (int)d1_cnt == d2_dst_pitch) {
         dma_2d_loc2sys_dyn_ch(ch,
                               src, dst, d1_cnt * d2_cnt, d3_src_pitch, d3_dst_pitch, d3_cnt);
-    } else if (d2_cnt * d2_src_pitch == d3_src_pitch
-               && d2_cnt * d2_dst_pitch == d3_dst_pitch) {
+    } else if ((int)d2_cnt * d2_src_pitch == d3_src_pitch
+               && (int)d2_cnt * d2_dst_pitch == d3_dst_pitch) {
         dma_2d_loc2sys_dyn_ch(ch,
                               src, dst, d1_cnt, d2_src_pitch, d2_dst_pitch, d2_cnt * d3_cnt);
     } else {
@@ -843,12 +843,12 @@ void dma_3d_sys2loc_dyn_ch(int ch,
                            int d3_src_pitch, int d3_dst_pitch, size_t d3_cnt)
 {
     /* Fold dimensions if possible */
-    if (d1_cnt == d2_src_pitch
-        && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch
+        && (int)d1_cnt == d2_dst_pitch) {
         dma_2d_sys2loc_dyn_ch(ch,
                               src, dst, d1_cnt * d2_cnt, d3_src_pitch, d3_dst_pitch, d3_cnt);
-    } else if (d2_cnt * d2_src_pitch == d3_src_pitch
-               && d2_cnt * d2_dst_pitch == d3_dst_pitch) {
+    } else if ((int)d2_cnt * d2_src_pitch == d3_src_pitch
+               && (int)d2_cnt * d2_dst_pitch == d3_dst_pitch) {
         dma_2d_sys2loc_dyn_ch(ch,
                               src, dst, d1_cnt, d2_src_pitch, d2_dst_pitch, d2_cnt * d3_cnt);
     } else {
@@ -1023,7 +1023,7 @@ void dma_2d_loc2sys_dyn_straddles_ch(int ch,
                                      void *src, void *dst, size_t d1_cnt, int d2_src_pitch, int d2_dst_pitch, size_t d2_cnt)
 {
     unsigned added;
-    if (d1_cnt == d2_src_pitch && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch && (int)d1_cnt == d2_dst_pitch) {
         added = add_2d_desc(ch,
                             src, dst, d1_cnt * d2_cnt, d1_cnt * d2_cnt, d1_cnt * d2_cnt, 1,
                             /*added*/ 0, /*isSys2Loc*/ 0, /*straddles*/ 1);
@@ -1041,7 +1041,7 @@ void dma_2d_sys2loc_dyn_straddles_ch(int ch,
                                      void *src, void *dst, size_t d1_cnt, int d2_src_pitch, int d2_dst_pitch, size_t d2_cnt)
 {
     unsigned added;
-    if (d1_cnt == d2_src_pitch && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch && (int)d1_cnt == d2_dst_pitch) {
         added = add_2d_desc(ch,
                             src, dst, d1_cnt * d2_cnt, d1_cnt * d2_cnt, d1_cnt * d2_cnt, 1,
                             /*added*/ 0, /*isSys2Loc*/ 1, /*straddles*/ 1);
@@ -1133,12 +1133,12 @@ void dma_3d_loc2sys_dyn_straddles_ch(int ch,
                                      int d3_src_pitch, int d3_dst_pitch, size_t d3_cnt)
 {
     /* Fold dimensions if possible */
-    if (d1_cnt == d2_src_pitch
-        && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch
+        && (int)d1_cnt == d2_dst_pitch) {
         dma_2d_loc2sys_dyn_straddles_ch(ch,
                                         src, dst, d1_cnt * d2_cnt, d3_src_pitch, d3_dst_pitch, d3_cnt);
-    } else if (d2_cnt * d2_src_pitch == d3_src_pitch
-               && d2_cnt * d2_dst_pitch == d3_dst_pitch) {
+    } else if ((int)d2_cnt * d2_src_pitch == d3_src_pitch
+               && (int)d2_cnt * d2_dst_pitch == d3_dst_pitch) {
         dma_2d_loc2sys_dyn_straddles_ch(ch,
                                         src, dst, d1_cnt, d2_src_pitch, d2_dst_pitch, d2_cnt * d3_cnt);
     } else {
@@ -1155,12 +1155,12 @@ void dma_3d_sys2loc_dyn_straddles_ch(int ch,
                                      int d3_src_pitch, int d3_dst_pitch, size_t d3_cnt)
 {
     /* Fold dimensions if possible */
-    if (d1_cnt == d2_src_pitch
-        && d1_cnt == d2_dst_pitch) {
+    if ((int)d1_cnt == d2_src_pitch
+        && (int)d1_cnt == d2_dst_pitch) {
         dma_2d_sys2loc_dyn_straddles_ch(ch,
                                         src, dst, d1_cnt * d2_cnt, d3_src_pitch, d3_dst_pitch, d3_cnt);
-    } else if (d2_cnt * d2_src_pitch == d3_src_pitch
-               && d2_cnt * d2_dst_pitch == d3_dst_pitch) {
+    } else if ((int)d2_cnt * d2_src_pitch == d3_src_pitch
+               && (int)d2_cnt * d2_dst_pitch == d3_dst_pitch) {
         dma_2d_sys2loc_dyn_straddles_ch(ch,
                                         src, dst, d1_cnt, d2_src_pitch, d2_dst_pitch, d2_cnt * d3_cnt);
     } else {
